Updated macro-related highlighting
Double-quote strings are now the only things that count as strings, and
define only counts as a macro definition if followed by a parenthesis. `
and ' now count as operators. I'm guessing the inconvenience of dealing
with non-highlighting in macro definitions would be worse than having a
non-matched single-quote string (which I've never seen in SPARC).